T #2 Switched Huppy to use Hosted Service
